Remove display of error code
Occasionally caused out of memory issue with the display. Instead of
printing the error text and error code, "of memory" was shown. This
appears to occur when an error happens during the HTTP call.
"of memory" is assumed to be "out of memory", but where it comes
from is unclear.